Refactors bin script into module
Moved most of the code from bin/aws-upload that was responsible for
the s3 upload to it's own file in the lib directory. I saw the
need for using this code in more than one place so the lib/s3-upload
module can now be used from more than one place in our app.
The bin/aws-upload script still works and just implements the lib once.